fossil running

+++ sys/src/cmd/fossil/pack.c
@@ -111,7 +111,7 @@ globalToLocal(uint8_t score[VtScoreSize])
 
 	for(i=0; i<VtScoreSize-4; i++)
 		if(score[i] != 0)
-			return (int64_t)NilBlock;
+			return NilBlock;
 
 	return U32GET(score+VtScoreSize-4);
 }
